Moved newImoraData to feed screen, added props to feedbox (#26)
