+++ Mobile/src/App.js
@@ -9,7 +9,7 @@ export default function App() {
   return (
     <>
       <StatusBar backgroundColor={'#4D2C91'} />
-      <ResidenceAddNavigation />
+      <MainNavigation />
     </>
   );
 }
